ClickPos/requirements.txt
pillow==10.2.0
pygame==2.5.2
